[Refactor] MemberController 불필요한 매개변수 제거, Oauthsuccess핸들러 닉네임 부분 수정

See merge request s08-webmobile2-sub2/S08P12B111!26

+++ back/src/main/java/com/udteam/miristock/dto/ArticleRequestDto.java
@@ -11,7 +11,7 @@ import lombok.*;
 @NoArgsConstructor
 public class ArticleRequestDto {
 
-    private Long articleNo; // 글 번호
+    private Integer articleNo; // 글 번호
     private String memberNickname; // 작성자 닉네임
     private String articleTitle; // 제목
     private String articleContent; // 내용
