+++ back/src/main/java/com/udteam/miristock/repository/ArticleRepository.java
@@ -3,6 +3,6 @@ package com.udteam.miristock.repository;
 import com.udteam.miristock.entity.ArticleEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArticleRepository extends JpaRepository<ArticleEntity, Long> {
+public interface ArticleRepository extends JpaRepository<ArticleEntity, Integer> {
 
 }
