+++ back/src/main/java/com/udteam/miristock/repository/MemberRepository.java
@@ -4,11 +4,11 @@ import com.udteam.miristock.entity.MemberEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import javax.transaction.Transactional;
-public interface MemberRepository extends JpaRepository<MemberEntity,Long>{
+public interface MemberRepository extends JpaRepository<MemberEntity, Integer>{
     MemberEntity findByMemberEmail(String email);
 
     @Transactional
-    int deleteByMemberEmail(String email);
+    Integer deleteByMemberEmail(String email);
     String findRefreshtokenByMemberEmail(String email);
 
 }
